requirements.txt

transformers
opencv-python
timm
einops
datasets
rouge_score
omegaconf
ftfy
pycocotools
pycocoevalcap
albumentations
pudb
imageio
imageio-ffmpeg
clean-fid
h5py
pillow
pytorch-lightning==1.5.9
tensorboardX==2.4.1
test-tube>=0.7.5
streamlit>=0.73.1
setuptools==59.5.0
torch-fidelity==0.3.0
torchmetrics==0.6.0
kornia==0.6
chardet
cchardet